Refactor TerrainSystem into focused components
Split the monolithic TerrainSystem (~1750 lines) into:
- TerrainHeightMap: Height data generation, GPU texture, CPU queries
- TerrainTextures: Albedo and grass far LOD textures
- TerrainCBT: Concurrent Binary Tree buffer management

TerrainSystem now uses composition to delegate to these classes while
maintaining its existing public API. This improves maintainability and
makes TerrainHeightMap reusable for physics/collision.
