+++ src/TerrainTextures.h
@@ -1,0 +1,53 @@
+#pragma once
+
+#include <vulkan/vulkan.h>
+#include <vk_mem_alloc.h>
+
+// Terrain textures - albedo and grass far LOD textures
+class TerrainTextures {
+public:
+    struct InitInfo {
+        VkDevice device;
+        VmaAllocator allocator;
+        VkQueue graphicsQueue;
+        VkCommandPool commandPool;
+    };
+
+    TerrainTextures() = default;
+    ~TerrainTextures() = default;
+
+    bool init(const InitInfo& info);
+    void destroy(VkDevice device, VmaAllocator allocator);
+
+    // Terrain albedo texture
+    VkImageView getAlbedoView() const { return albedoView; }
+    VkSampler getAlbedoSampler() const { return albedoSampler; }
+
+    // Grass far LOD texture (for terrain blending at distance)
+    VkImageView getGrassFarLODView() const { return grassFarLODView; }
+    VkSampler getGrassFarLODSampler() const { return grassFarLODSampler; }
+
+private:
+    bool createAlbedoTexture();
+    bool createGrassFarLODTexture();
+    bool uploadImageData(VkImage image, const void* data, uint32_t width, uint32_t height,
+                         VkFormat format, uint32_t bytesPerPixel);
+
+    // Init params
+    VkDevice device = VK_NULL_HANDLE;
+    VmaAllocator allocator = VK_NULL_HANDLE;
+    VkQueue graphicsQueue = VK_NULL_HANDLE;
+    VkCommandPool commandPool = VK_NULL_HANDLE;
+
+    // Terrain albedo texture
+    VkImage albedoImage = VK_NULL_HANDLE;
+    VmaAllocation albedoAllocation = VK_NULL_HANDLE;
+    VkImageView albedoView = VK_NULL_HANDLE;
+    VkSampler albedoSampler = VK_NULL_HANDLE;
+
+    // Grass far LOD texture
+    VkImage grassFarLODImage = VK_NULL_HANDLE;
+    VmaAllocation grassFarLODAllocation = VK_NULL_HANDLE;
+    VkImageView grassFarLODView = VK_NULL_HANDLE;
+    VkSampler grassFarLODSampler = VK_NULL_HANDLE;
+};
